# helpers/swagger.yaml
openapi: 3.0.0
info:
  title: The System MySQL Database
  description: The System MySQL Database
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /cases:
    get:
      summary: Get a list of all cases
      description: Retrieve list of all cases.
      operationId: getAllCases
      security:
        - bearerAuth: []
      responses:
        '200':
          description: An array of all cases
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 123
                    caseNumber:
                      type: integer
                      example: 123
                    initialAssignment:
                      type: string
                      example: 'mickey@example.com'
                    currentAssignment:
                      type: string
                      example: 'mickey@example.com'
                    nextVisitDateTime:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                      description: 'Date the case must be worked on again.'
                    caseNotes:
                      type: string
                      example: 'Notes of the case provided by anyone'
                    kamNotes:
                      type: string
                      example: 'Notes of the case provided only by the KAM'
                    currentStatus:
                      type: string
                      example: 'Open'
                    resolution:
                      type: string
                      example: 'What was down to resolve the case - provided by a dropdown on the front end'
                    lockedDatetime:
                      type: string
                      minLength: 8
                      maxLength: 8
                      format: date
                      pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                      example: 20190816
                      description: 'Date the case was locked by opening it for action. The server uses the date to automatically unlock it if not resolved by the user'
                    createdDate:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    createdBy:
                      type: string
                      example: 'mickey@example.com'
                    updatedDate:
                      type: string
                      example: '2020-05-08T03:11:21.553Z'
                    updatedBy:
                      type: string
                      example: 'mickey@example.com'
                    f_accountNumber:
                      type: string
                      example: 'AAA101'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    post:
      summary: Create a new case
      description: Restricted to admin users.
      operationId: createCase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  caseNumber:
                    type: integer
                    example: 123
                  initialAssignment:
                    type: string
                    example: 'mickey@example.com'
                  currentAssignment:
                    type: string
                    example: 'mickey@example.com'
                  nextVisitDateTime:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the case must be worked on again.'
                  caseNotes:
                    type: string
                    example: 'Notes of the case provided by anyone'
                  kamNotes:
                    type: string
                    example: 'Notes of the case provided only by the KAM'
                  currentStatus:
                    type: string
                    example: 'Open'
                  resolution:
                    type: string
                    example: 'What was down to resolve the case - provided by a dropdown on the front end'
                  lockedDatetime:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the case was locked by opening it for action. The server uses the date to automatically unlock it if not resolved by the user'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_accountNumber:
                    type: string
                    example: 'AAA101'
              required:
                - createdDate
                - createdBy
                - currentAssignment
                - currentStatus
                - f_accountNumber
      responses:
        '200':
          description: Case created successfully. The id of the new case is returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
        '400':
          description: Case id already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Case id '123' already exists"
        '401':
          $ref: '#/components/responses/UnauthorisedError'
  /cases/{id}:
    parameters:
      - in: path
        name: id
        description: Case id
        required: true
        example: 123
        schema:
          type: integer
    get:
      summary: Get a single case by id
      description: Retrieve a single case by id.
      operationId: getCaseById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Details of the specified case
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  caseNumber:
                    type: integer
                    example: 123
                  initialAssignment:
                    type: string
                    example: 'mickey@example.com'
                  currentAssignment:
                    type: string
                    example: 'mickey@example.com'
                  nextVisitDateTime:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the case must be worked on again.'
                  caseNotes:
                    type: string
                    example: 'Notes of the case provided by anyone'
                  kamNotes:
                    type: string
                    example: 'Notes of the case provided only by the KAM'
                  currentStatus:
                    type: string
                    example: 'Open'
                  resolution:
                    type: string
                    example: 'What was down to resolve the case - provided by a dropdown on the front end'
                  lockedDatetime:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the case was locked by opening it for action. The server uses the date to automatically unlock it if not resolved by the user'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_accountNumber:
                    type: string
                    example: 'AAA101'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    put:
      summary: Update a case
      description: Update details of a case.
      operationId: updateCase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 123
                caseNumber:
                  type: integer
                  example: 123
                initialAssignment:
                  type: string
                  example: 'mickey@example.com'
                currentAssignment:
                  type: string
                  example: 'mickey@example.com'
                nextVisitDateTime:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                  description: 'Date the case must be worked on again.'
                caseNotes:
                  type: string
                  example: 'Notes of the case provided by anyone'
                kamNotes:
                  type: string
                  example: 'Notes of the case provided only by the KAM'
                currentStatus:
                  type: string
                  example: 'Open'
                resolution:
                  type: string
                  example: 'What was down to resolve the case - provided by a dropdown on the front end'
                lockedDatetime:
                  type: string
                  minLength: 8
                  maxLength: 8
                  format: date
                  pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                  example: 20190816
                  description: 'Date the case was locked by opening it for action. The server uses the date to automatically unlock it if not resolved by the user'
                createdDate:
                  type: string
                  example: '2020-05-05T09:12:57.848Z'
                createdBy:
                  type: string
                  example: 'mickey@example.com'
                updatedDate:
                  type: string
                  example: '2020-05-08T03:11:21.553Z'
                updatedBy:
                  type: string
                  example: 'mickey@example.com'
                f_accountNumber:
                  type: string
                  example: 'AAA101'
      responses:
        '200':
          description: Case updated successfully. The details of the updated case are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  caseNumber:
                    type: integer
                    example: 123
                  initialAssignment:
                    type: string
                    example: 'mickey@example.com'
                  currentAssignment:
                    type: string
                    example: 'mickey@example.com'
                  nextVisitDateTime:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the case must be worked on again.'
                  caseNotes:
                    type: string
                    example: 'Notes of the case provided by anyone'
                  kamNotes:
                    type: string
                    example: 'Notes of the case provided only by the KAM'
                  currentStatus:
                    type: string
                    example: 'Open'
                  resolution:
                    type: string
                    example: 'What was down to resolve the case - provided by a dropdown on the front end'
                  lockedDatetime:
                    type: string
                    minLength: 8
                    maxLength: 8
                    format: date
                    pattern: '^\d{4}(0[1-9]|1[012])(0[1-9]|[12][0-9]|3[01])$'
                    example: 20190816
                    description: 'Date the case was locked by opening it for action. The server uses the date to automatically unlock it if not resolved by the user'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  updatedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  updatedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_accountNumber:
                    type: string
                    example: 'AAA101'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    delete:
      summary: Delete a case
      description: Delete a case.
      operationId: deleteCase
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Case deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Case deleted successfully'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorisedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Unauthorised'
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Not Found'